backend/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
torch
transformers
sentence-transformers
scikit-learn
python-multipart
